requirements.txt
torch
numpy
pillow
pandas
opencv-python
torchvision
	
